update
git-svn-id: http://svn.tudos.org/repos/tudos/trunk@340 a704ac0b-3a55-4d43-a2a9-7be6f07c34fb

+++ dice/src/parser/Converter.cpp
@@ -232,7 +232,7 @@ long dice::parser::hex_to_long (char *s, IntType& t)
 		break;
 	case INT_INT:
 		{
-			long n = strtol(s, 0, 16);
+			int n = strtol(s, 0, 16);
 			if (! (std::numeric_limits<int>::min() <= n &&
 					n <= std::numeric_limits<int>::max() &&
 					errno != ERANGE))
